requirements.txt
a2a-sdk>=0.3.0
fastapi>=0.116.1
mcp[cli]>=1.12.3
openai>=1.98.0
qdrant-client>=1.15.0
duckduckgo-search>=6.1.0
sentence-transformers>=3.0.0
spacy>=3.7.0
tika>=2.6.0
scikit-learn>=1.3.0
numpy>=1.24.0
pydantic>=2.0.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
